rename vrt.py to stack.py
