+++ tests/test_sequential.py
@@ -3,7 +3,7 @@ import numpy.testing as npt
 import pytest
 from osgeo import gdal
 
-from dolphin import io, sequential, vrt
+from dolphin import io, sequential, stack
 from dolphin.phase_link import mle, simulate
 from dolphin.utils import gpu_is_available
 
@@ -17,7 +17,7 @@ pytestmark = pytest.mark.filterwarnings(
 
 
 def test_setup_output_folder(tmpdir, tiled_file_list):
-    vrt_stack = vrt.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
+    vrt_stack = stack.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
     out_file_list = sequential._setup_output_folder(
         vrt_stack, driver="GTiff", dtype=np.complex64
     )
@@ -46,7 +46,7 @@ def test_setup_output_folder(tmpdir, tiled_file_list):
     "strides", [{"x": 1, "y": 1}, {"x": 1, "y": 2}, {"x": 2, "y": 3}, {"x": 4, "y": 2}]
 )
 def test_setup_output_folder_strided(tmpdir, tiled_file_list, strides):
-    vrt_stack = vrt.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
+    vrt_stack = stack.VRTStack(tiled_file_list, outfile=tmpdir / "stack.vrt")
 
     out_file_list = sequential._setup_output_folder(
         vrt_stack, driver="GTiff", dtype=np.complex64, strides=strides
@@ -68,7 +68,7 @@ def test_setup_output_folder_strided(tmpdir, tiled_file_list, strides):
 def test_sequential_gtiff(tmp_path, slc_file_list, gpu_enabled):
     """Run through the sequential estimation with a GeoTIFF stack."""
     vrt_file = tmp_path / "slc_stack.vrt"
-    vrt_stack = vrt.VRTStack(slc_file_list, outfile=vrt_file)
+    vrt_stack = stack.VRTStack(slc_file_list, outfile=vrt_file)
     _, rows, cols = vrt_stack.shape
 
     half_window = {"x": cols // 2, "y": rows // 2}
@@ -114,7 +114,7 @@ def test_sequential_gtiff(tmp_path, slc_file_list, gpu_enabled):
 def test_sequential_nc(tmp_path, slc_file_list_nc, half_window, strides):
     """Check various strides/windows/ministacks with a NetCDF input stack."""
     vrt_file = tmp_path / "slc_stack.vrt"
-    _ = vrt.VRTStack(slc_file_list_nc, outfile=vrt_file, subdataset="data")
+    _ = stack.VRTStack(slc_file_list_nc, outfile=vrt_file, subdataset="data")
 
     output_folder = tmp_path / "sequential"
     sequential.run_evd_sequential(
@@ -135,7 +135,9 @@ def test_sequential_ministack_sizes(tmp_path, slc_file_list_nc, ministack_size):
     """Check various strides/windows/ministacks with a NetCDF input stack."""
     vrt_file = tmp_path / "slc_stack.vrt"
     # Make it not a round number to test
-    vrt_stack = vrt.VRTStack(slc_file_list_nc[:21], outfile=vrt_file, subdataset="data")
+    vrt_stack = stack.VRTStack(
+        slc_file_list_nc[:21], outfile=vrt_file, subdataset="data"
+    )
     _, rows, cols = vrt_stack.shape
 
     output_folder = tmp_path / "sequential"
